Service Interface

+++ src/main/java/com/alaia/pharmX/repositories/ContactRepository.java
@@ -3,5 +3,5 @@ package com.alaia.pharmX.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.alaia.pharmX.models.Contact;
 
-public interface ContactRepository extends JpaRepository<Contact, Integer> {
+public interface ContactRepository extends JpaRepository<Contact, Long> {
 }
