+++ src/main/java/com/alaia/pharmX/repositories/OrderLineRepository.java
@@ -3,5 +3,5 @@ package com.alaia.pharmX.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.alaia.pharmX.models.OrderLine;
 
-public interface OrderLineRepository extends JpaRepository<OrderLine, Integer> {
+public interface OrderLineRepository extends JpaRepository<OrderLine, Long> {
 }
