+++ src/main/java/com/alaia/pharmX/repositories/SectionRepository.java
@@ -3,5 +3,5 @@ package com.alaia.pharmX.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.alaia.pharmX.models.Section;
 
-public interface SectionRepository extends JpaRepository<Section, Integer> {
+public interface SectionRepository extends JpaRepository<Section, Long> {
 }
