+++ src/main/java/com/alaia/pharmX/repositories/SlotRepository.java
@@ -3,5 +3,5 @@ package com.alaia.pharmX.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.alaia.pharmX.models.Slot;
 
-public interface SlotRepository extends JpaRepository<Slot, Integer> {
+public interface SlotRepository extends JpaRepository<Slot, Long> {
 }
